requirements.txt
pandas # libreria para los dataframes
nump
Faker # libreria para numeros aleatorios

# Para conectar Python con PostgreSQL
sqlalchemy      
psycopg2-binary 